+++ web/js/jquery.gerbera.clients.js
@@ -37,10 +37,11 @@ $.widget('grb.clients', {
       matchType: 'Match Type',
       match: 'Match Pattern',
       clientType: 'Client Type',
-      flags: 'Quirk Flags'
+      flags: 'Client Flags',
+      empty: '',
     };
-    const clientProps = ['ip', 'host', 'group', 'name', 'userAgent', 'matchType', 'match', 'clientType', 'time', 'last', 'flags' ];
-    this.buildTable (table, this.options.data, clientHeadings, clientProps, 'Clients', clientProps.length);
+    const clientProps = [['ip', 'host', 'group'], 'userAgent', ['name', 'matchType', 'match'], 'clientType', ['time', 'last'], 'flags'];
+    this.buildTable(table, this.options.data, clientHeadings, clientProps, 'Clients', clientProps.length, this.options.onDelete);
 
     const groupHeadings = {
       name: 'Group',
@@ -50,8 +51,8 @@ $.widget('grb.clients', {
       last: 'Last Action',
       empty: '',
     };
-    const groupProps = ['empty', 'empty', 'name', 'empty', 'empty', 'count', 'playCount', 'bookmarks', 'empty', 'last', 'empty' ];
-    this.buildTable (table, this.options.groups, groupHeadings, groupProps, 'Groups', clientProps.length);
+    const groupProps = ['name', 'count', 'playCount', 'bookmarks', 'last', 'empty', 'empty'];
+    this.buildTable(table, this.options.groups, groupHeadings, groupProps, 'Groups', clientProps.length);
     this.element.append(table);
     this.element.addClass('with-data');
   },
@@ -62,21 +63,32 @@ $.widget('grb.clients', {
     this.element.removeClass('with-data');
   },
 
-  buildTable: function (table, data, headings, props, caption, size) {
+  buildTable: function (table, data, headings, props, caption, size, onDelete) {
     const tbody = $('<tbody></tbody>');
     const thead = $('<thead></thead>');
     let row, content, text;
 
     if (data && data.length > 0) {
       row = $('<tr></tr>');
-      props.forEach( function(p) {
-          content = $('<th></th>');
+      props.forEach(function (prop) {
+        content = $('<th></th>');
+        const pL = (Array.isArray(prop)) ? prop : [prop];
+        pL.forEach(function (p, idx) {
+          if (idx > 0)
+            $('<br><hr>').appendTo(content);
           text = $('<span></span>');
           text.text(headings[p]).appendTo(content);
           content.addClass('grb-client-' + p);
-
-          row.append(content);
+        });
+        row.append(content);
       });
+      if (onDelete) {
+        content = $('<th></th>');
+        const deleteIcon = $('<span>Action</span>');
+        deleteIcon.addClass('grb-client-delete');
+        deleteIcon.appendTo(content);
+        row.append(content);
+      }
 
       row.addClass('grb-client');
       thead.append(row);
@@ -85,14 +97,30 @@ $.widget('grb.clients', {
         const item = data[i];
         row = $('<tr></tr>');
 
-        props.forEach( function(p) {
-            content = $('<td></td>');
+        props.forEach(function (prop) {
+          content = $('<td></td>');
+          const pL = (Array.isArray(prop)) ? prop : [prop];
+          pL.forEach(function (p, idx) {
+            if (idx > 0)
+              $('<br><hr>').appendTo(content);
             text = $('<span></span>');
             text.text(item[p]).appendTo(content);
             content.addClass('grb-client-' + p);
-            row.append(content);
+          });
+          row.append(content);
         });
-
+        if (onDelete) {
+          content = $('<td></td>');
+          const deleteIcon = $('<span></span>');
+          deleteIcon.prop('title', 'Delete client');
+          deleteIcon.addClass('grb-client-delete fa fa-trash-o');
+          deleteIcon.appendTo(content);
+          row.append(content);
+          deleteIcon.click(item, function (event) {
+            row.remove();
+            onDelete(event.data);
+          });
+        }
         row.addClass('grb-client');
         tbody.append(row);
       }